Extract JPackage plugin into a standalone project
It is big, and having it in buildSrc is inconvenient - any change to it
needlessly invalidates cached java compilation tasks (because of
the classpath change).

Issue: n/a
