+++ gradle/plugins/jpackage/build.gradle.kts
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 the Andlogview authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    `kotlin-dsl`
+}
+
+repositories {
+    gradlePluginPortal()
+    mavenCentral()
+}
+
+dependencies {
+    implementation(libs.build.jlink.plugin)
+}
+
+description = "Convenience Wrapper for the JPackage Tool"
+
+gradlePlugin {
+    plugins {
+        create("jpackage") {
+            id = "name.mlopatkin.gradleplugins.jpackage"
+            implementationClass = "name.mlopatkin.gradleplugins.jpackage.JpackagePlugin"
+        }
+    }
+}
