+++ settings.gradle.kts
@@ -16,6 +16,7 @@
 
 pluginManagement {
     includeBuild("gradle/plugins/freemarker")
+    includeBuild("gradle/plugins/jpackage")
 }
 
 dependencyResolutionManagement {
